added security for the other endpoints, added hidden functionality for projects sprints and tasks, Defined the relationship between projects and employees and teams better

+++ src/main/java/com/example/manager/client/project/ProjectClient.java
@@ -1,6 +1,7 @@
 package com.example.manager.client.project;
 
 import com.example.manager.core.application.project.ProjectInterface;
+import com.example.manager.core.application.project.ProjectService;
 import com.example.manager.core.application.repositories.ProjectRepository;
 import com.example.manager.core.domain.Project;
 import org.springframework.stereotype.Service;
@@ -12,13 +13,16 @@ import java.util.Optional;
 public class ProjectClient implements ProjectInterface {
 
     private final ProjectRepository projectRepository;
+    private final ProjectService projectService;
 
-    public ProjectClient(ProjectRepository projectRepository) {
+    public ProjectClient(ProjectRepository projectRepository, ProjectService projectService) {
         this.projectRepository = projectRepository;
+        this.projectService = projectService;
     }
 
     @Override
     public Project saveProject(Project project) {
+        projectService.addProjectToMembersOfTeam(project);
         return projectRepository.save(project);
     }
 
@@ -33,13 +37,28 @@ public class ProjectClient implements ProjectInterface {
     }
 
     @Override
-    public Optional<Project> updateProjectById(Project project, Long projectId) {
-        projectRepository.deleteById(projectId);
+    public Optional<Project> updateProjectById(Project project, Long uid) {
+        Optional<Project> projectBeforeUpdate = projectRepository.findById(uid);
+        if(projectBeforeUpdate.isEmpty()){
+            return Optional.empty();
+        }
+        if(!projectBeforeUpdate.get().getTeam().equals(project.getTeam())){
+            projectService.deleteProjectFromMembersOfTeam(projectBeforeUpdate.get());
+            projectService.addProjectToMembersOfTeam(project);
+            projectService.deleteAllSprintsWithProject(projectBeforeUpdate.get());
+        }
+        projectRepository.deleteById(uid);
         return Optional.of(projectRepository.save(project));
     }
 
     @Override
     public void deleteProjectById(Long uid) {
+        Optional<Project> projectBeforeDelete = projectRepository.findById(uid);
+        if(projectBeforeDelete.isEmpty()){
+            return;
+        }
+        projectService.deleteAllSprintsWithProject(projectBeforeDelete.get());
+        projectService.deleteProjectFromMembersOfTeam(projectBeforeDelete.get());
         projectRepository.deleteById(uid);
     }
 }
